Update Gamecock.py
added a specificality to the searching of equity swaps.
if quotes are used, it will be specific to the term, negating all other terms which are not quoted. aka "gme" will not include gme.n in the results.
but if no quotes are used, it will include all entries which contain the term, regardless of the terms placement in the records.

this should allow for specificity in the users choice of terms.